split sorter into sync vs async
